Move out weight conversion functions into individual separate files. Plus a unit test to ensure the pattern followed is accessible.
