+++ transformer_lens/loading_from_pretrained.py
@@ -7,16 +7,34 @@ import dataclasses
 import logging
 import os
 import re
-from typing import Dict, Optional, Union, cast
+from typing import Dict, Optional, Union
 
-import einops
 import torch
 from huggingface_hub import HfApi
 from transformers import AutoConfig, AutoModelForCausalLM, BertForPreTraining
 
 import transformer_lens.utils as utils
 from transformer_lens.HookedTransformerConfig import HookedTransformerConfig
-from transformer_lens.pretrained.weight_conversions import convert_mixtral_weights
+from transformer_lens.pretrained.weight_conversions import (
+    convert_bert_weights,
+    convert_bloom_weights,
+    convert_coder_weights,
+    convert_gemma_weights,
+    convert_gpt2_weights,
+    convert_gptj_weights,
+    convert_llama_weights,
+    convert_mingpt_weights,
+    convert_mistral_weights,
+    convert_mixtral_weights,
+    convert_neel_solu_old_weights,
+    convert_neo_weights,
+    convert_neox_weights,
+    convert_opt_weights,
+    convert_phi3_weights,
+    convert_phi_weights,
+    convert_qwen2_weights,
+    convert_qwen_weights,
+)
 
 OFFICIAL_MODEL_NAMES = [
     "gpt2",
@@ -1569,1104 +1587,6 @@ def fill_missing_keys(model, state_dict):
                 )
             )
         state_dict[key] = default_state_dict[key]
-    return state_dict
-
-
-# Convert state dicts
-def convert_gpt2_weights(gpt2, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = gpt2.transformer.wte.weight
-    state_dict["pos_embed.W_pos"] = gpt2.transformer.wpe.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = gpt2.transformer.h[l].ln_1.weight
-        state_dict[f"blocks.{l}.ln1.b"] = gpt2.transformer.h[l].ln_1.bias
-
-        # In GPT-2, q,k,v are produced by one big linear map, whose output is
-        # concat([q, k, v])
-        W = gpt2.transformer.h[l].attn.c_attn.weight
-        W_Q, W_K, W_V = torch.tensor_split(W, 3, dim=1)
-        W_Q = einops.rearrange(W_Q, "m (i h)->i m h", i=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "m (i h)->i m h", i=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "m (i h)->i m h", i=cfg.n_heads)
-
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        qkv_bias = gpt2.transformer.h[l].attn.c_attn.bias
-        qkv_bias = einops.rearrange(
-            qkv_bias,
-            "(qkv index head)->qkv index head",
-            qkv=3,
-            index=cfg.n_heads,
-            head=cfg.d_head,
-        )
-        state_dict[f"blocks.{l}.attn.b_Q"] = qkv_bias[0]
-        state_dict[f"blocks.{l}.attn.b_K"] = qkv_bias[1]
-        state_dict[f"blocks.{l}.attn.b_V"] = qkv_bias[2]
-
-        W_O = gpt2.transformer.h[l].attn.c_proj.weight
-        W_O = einops.rearrange(W_O, "(i h) m->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = gpt2.transformer.h[l].attn.c_proj.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = gpt2.transformer.h[l].ln_2.weight
-        state_dict[f"blocks.{l}.ln2.b"] = gpt2.transformer.h[l].ln_2.bias
-
-        W_in = gpt2.transformer.h[l].mlp.c_fc.weight
-        state_dict[f"blocks.{l}.mlp.W_in"] = W_in
-        state_dict[f"blocks.{l}.mlp.b_in"] = gpt2.transformer.h[l].mlp.c_fc.bias
-
-        W_out = gpt2.transformer.h[l].mlp.c_proj.weight
-        state_dict[f"blocks.{l}.mlp.W_out"] = W_out
-        state_dict[f"blocks.{l}.mlp.b_out"] = gpt2.transformer.h[l].mlp.c_proj.bias
-    state_dict["unembed.W_U"] = gpt2.lm_head.weight.T
-
-    state_dict["ln_final.w"] = gpt2.transformer.ln_f.weight
-    state_dict["ln_final.b"] = gpt2.transformer.ln_f.bias
-    return state_dict
-
-
-def convert_neo_weights(neo, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = neo.transformer.wte.weight
-    state_dict["pos_embed.W_pos"] = neo.transformer.wpe.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = neo.transformer.h[l].ln_1.weight
-        state_dict[f"blocks.{l}.ln1.b"] = neo.transformer.h[l].ln_1.bias
-
-        W_Q = neo.transformer.h[l].attn.attention.q_proj.weight
-        W_K = neo.transformer.h[l].attn.attention.k_proj.weight
-        W_V = neo.transformer.h[l].attn.attention.v_proj.weight
-        W_Q = einops.rearrange(W_Q, "(i h) m->i m h", i=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(i h) m->i m h", i=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(i h) m->i m h", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.b_K"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.b_V"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-
-        W_O = neo.transformer.h[l].attn.attention.out_proj.weight
-        W_O = einops.rearrange(W_O, "m (i h)->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = neo.transformer.h[l].attn.attention.out_proj.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = neo.transformer.h[l].ln_2.weight
-        state_dict[f"blocks.{l}.ln2.b"] = neo.transformer.h[l].ln_2.bias
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = neo.transformer.h[l].mlp.c_fc.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = neo.transformer.h[l].mlp.c_fc.bias
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = neo.transformer.h[l].mlp.c_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = neo.transformer.h[l].mlp.c_proj.bias
-    state_dict["ln_final.w"] = neo.transformer.ln_f.weight
-    state_dict["ln_final.b"] = neo.transformer.ln_f.bias
-
-    state_dict["unembed.W_U"] = neo.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-    return state_dict
-
-
-def convert_gptj_weights(gptj, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = gptj.transformer.wte.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = gptj.transformer.h[l].ln_1.weight
-        state_dict[f"blocks.{l}.ln1.b"] = gptj.transformer.h[l].ln_1.bias
-
-        W_Q = gptj.transformer.h[l].attn.q_proj.weight
-        W_K = gptj.transformer.h[l].attn.k_proj.weight
-        W_V = gptj.transformer.h[l].attn.v_proj.weight
-        W_Q = einops.rearrange(W_Q, "(i h) m->i m h", i=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(i h) m->i m h", i=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(i h) m->i m h", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.b_K"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.b_V"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-
-        W_O = gptj.transformer.h[l].attn.out_proj.weight
-        W_O = einops.rearrange(W_O, "m (i h)->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        # Layer Norm 1 and 2 are tied.
-        state_dict[f"blocks.{l}.ln2.w"] = state_dict[f"blocks.{l}.ln1.w"]
-        state_dict[f"blocks.{l}.ln2.b"] = state_dict[f"blocks.{l}.ln1.b"]
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = gptj.transformer.h[l].mlp.fc_in.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = gptj.transformer.h[l].mlp.fc_in.bias
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = gptj.transformer.h[l].mlp.fc_out.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = gptj.transformer.h[l].mlp.fc_out.bias
-    state_dict["ln_final.w"] = gptj.transformer.ln_f.weight
-    state_dict["ln_final.b"] = gptj.transformer.ln_f.bias
-
-    state_dict["unembed.W_U"] = gptj.lm_head.weight.T
-    # Contains a bias, for some reason?
-    state_dict["unembed.b_U"] = gptj.lm_head.bias
-    return state_dict
-
-
-def convert_neox_weights(neox, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = neox.gpt_neox.embed_in.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = neox.gpt_neox.layers[l].input_layernorm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = neox.gpt_neox.layers[l].input_layernorm.bias
-
-        # For some inexplicable reason, NeoX both uses the concatenated QKV
-        # matmul of GPT-2 (afaict this has a neglible performance impact) AND
-        # has the flattened axis in the DIFFERENT order of (head_index qkv
-        # d_head) - this took me an hour to debug...
-        W = neox.gpt_neox.layers[l].attention.query_key_value.weight
-        W = einops.rearrange(W, "(i qkv h) m->qkv i m h", i=cfg.n_heads, qkv=3)
-
-        # Fold in layer norm weights
-        state_dict[f"blocks.{l}.attn.W_Q"] = W[0]
-        state_dict[f"blocks.{l}.attn.W_K"] = W[1]
-        state_dict[f"blocks.{l}.attn.W_V"] = W[2]
-
-        qkv_bias = neox.gpt_neox.layers[l].attention.query_key_value.bias
-        qkv_bias = einops.rearrange(
-            qkv_bias,
-            "(index qkv head)->qkv index head",
-            qkv=3,
-            index=cfg.n_heads,
-            head=cfg.d_head,
-        )
-        # Fold in layer norm biases
-        state_dict[f"blocks.{l}.attn.b_Q"] = qkv_bias[0]
-        state_dict[f"blocks.{l}.attn.b_K"] = qkv_bias[1]
-        state_dict[f"blocks.{l}.attn.b_V"] = qkv_bias[2]
-
-        W_O = neox.gpt_neox.layers[l].attention.dense.weight
-        W_O = einops.rearrange(W_O, "m (i h)->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = neox.gpt_neox.layers[l].attention.dense.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = neox.gpt_neox.layers[l].post_attention_layernorm.weight
-        state_dict[f"blocks.{l}.ln2.b"] = neox.gpt_neox.layers[l].post_attention_layernorm.bias
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = neox.gpt_neox.layers[l].mlp.dense_h_to_4h.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = neox.gpt_neox.layers[l].mlp.dense_h_to_4h.bias
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = neox.gpt_neox.layers[l].mlp.dense_4h_to_h.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = neox.gpt_neox.layers[l].mlp.dense_4h_to_h.bias
-    state_dict["ln_final.w"] = neox.gpt_neox.final_layer_norm.weight
-    state_dict["ln_final.b"] = neox.gpt_neox.final_layer_norm.bias
-
-    state_dict["unembed.W_U"] = neox.embed_out.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-    return state_dict
-
-
-def convert_llama_weights(llama, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = llama.model.embed_tokens.weight
-
-    # Some models with the Llama architecture use Grouped Query Attention, and so for these we need to modify
-    # the state dict keys for the K/V attention weight/biases, prepending "_" to the key names.
-    using_gqa = cfg.n_key_value_heads is not None
-    gqa_uscore = "_" if using_gqa else ""
-    # need a cast since MyPy isn't smart enough to realize that using_gqa implies n_key_value_heads is not None
-    n_kv_heads = cast(int, cfg.n_key_value_heads if using_gqa else cfg.n_heads)
-
-    # llama has no biases anywhere and deals with everything else roughly like
-    # GPTNeoX with different names
-
-    assert cfg.d_mlp is not None  # keep mypy happy
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = llama.model.layers[l].input_layernorm.weight
-
-        W_Q = llama.model.layers[l].self_attn.q_proj.weight
-        W_K = llama.model.layers[l].self_attn.k_proj.weight
-        W_V = llama.model.layers[l].self_attn.v_proj.weight
-
-        # in case of quantization,
-        # parameters should stay as bitsandbytes.nn.modules.Params4bit
-        if not cfg.load_in_4bit:
-            W_Q = einops.rearrange(W_Q, "(n h) m->n m h", n=cfg.n_heads)
-            W_K = einops.rearrange(W_K, "(n h) m->n m h", n=n_kv_heads)
-            W_V = einops.rearrange(W_V, "(n h) m->n m h", n=n_kv_heads)
-
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.{gqa_uscore}W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.{gqa_uscore}W_V"] = W_V
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(
-            cfg.n_heads, cfg.d_head, dtype=cfg.dtype, device=cfg.device
-        )
-        state_dict[f"blocks.{l}.attn.{gqa_uscore}b_K"] = torch.zeros(
-            n_kv_heads,
-            cfg.d_head,
-            dtype=cfg.dtype,
-            device=cfg.device,
-        )
-        state_dict[f"blocks.{l}.attn.{gqa_uscore}b_V"] = torch.zeros(
-            n_kv_heads,
-            cfg.d_head,
-            dtype=cfg.dtype,
-            device=cfg.device,
-        )
-
-        W_O = llama.model.layers[l].self_attn.o_proj.weight
-
-        if not cfg.load_in_4bit:
-            W_O = einops.rearrange(W_O, "m (n h)->n h m", n=cfg.n_heads)
-
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O.to(device=cfg.device)
-
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(
-            cfg.d_model, dtype=cfg.dtype, device=cfg.device
-        )
-
-        state_dict[f"blocks.{l}.ln2.w"] = llama.model.layers[l].post_attention_layernorm.weight
-
-        # in case of quantization,
-        # parameters should stay as bitsandbytes.nn.modules.Params4bit
-        if not cfg.load_in_4bit:
-            state_dict[f"blocks.{l}.mlp.W_in"] = llama.model.layers[l].mlp.up_proj.weight.T
-            state_dict[f"blocks.{l}.mlp.W_gate"] = llama.model.layers[l].mlp.gate_proj.weight.T
-            state_dict[f"blocks.{l}.mlp.W_out"] = llama.model.layers[l].mlp.down_proj.weight.T
-        else:
-            state_dict[f"blocks.{l}.mlp.W_in"] = llama.model.layers[l].mlp.up_proj.weight
-            state_dict[f"blocks.{l}.mlp.W_gate"] = llama.model.layers[l].mlp.gate_proj.weight
-            state_dict[f"blocks.{l}.mlp.W_out"] = llama.model.layers[l].mlp.down_proj.weight
-
-        state_dict[f"blocks.{l}.mlp.b_in"] = torch.zeros(
-            cfg.d_mlp, dtype=cfg.dtype, device=cfg.device
-        )
-        state_dict[f"blocks.{l}.mlp.b_out"] = torch.zeros(
-            cfg.d_model, dtype=cfg.dtype, device=cfg.device
-        )
-
-    state_dict["ln_final.w"] = llama.model.norm.weight
-
-    state_dict["unembed.W_U"] = llama.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype, device=cfg.device)
-
-    return state_dict
-
-
-def convert_qwen_weights(qwen, cfg: HookedTransformerConfig):
-    state_dict = {}
-    model = qwen.transformer
-    state_dict["embed.W_E"] = model.wte.weight
-
-    assert cfg.d_mlp is not None  # keep mypy happy
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = model.h[l].ln_1.weight
-
-        W_Q, W_K, W_V = model.h[l].attn.c_attn.weight.split(split_size=cfg.d_model, dim=0)
-        W_Q = einops.rearrange(W_Q, "(n h) m->n m h", n=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(n h) m->n m h", n=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(n h) m->n m h", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        b_Q, b_K, b_V = model.h[l].attn.c_attn.bias.split(split_size=cfg.d_model, dim=0)
-        b_Q = einops.rearrange(
-            b_Q,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-        b_K = einops.rearrange(
-            b_K,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-        b_V = einops.rearrange(
-            b_V,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-        state_dict[f"blocks.{l}.attn.b_Q"] = b_Q
-        state_dict[f"blocks.{l}.attn.b_K"] = b_K
-        state_dict[f"blocks.{l}.attn.b_V"] = b_V
-
-        W_O = model.h[l].attn.c_proj.weight
-        W_O = einops.rearrange(W_O, "m (n h)->n h m", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.ln2.w"] = model.h[l].ln_2.weight
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = model.h[l].mlp.w1.weight.T
-        state_dict[f"blocks.{l}.mlp.W_gate"] = model.h[l].mlp.w2.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = torch.zeros(cfg.d_mlp, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = model.h[l].mlp.c_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-    state_dict["ln_final.w"] = model.ln_f.weight
-
-    state_dict["unembed.W_U"] = qwen.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-    return state_dict
-
-
-def convert_qwen2_weights(qwen, cfg: HookedTransformerConfig):
-    # Note that this method is also applied for Qwen1.5 models, since they
-    # have architecture type Qwen2ForCausalLM.
-
-    state_dict = {}
-
-    state_dict["embed.W_E"] = qwen.model.embed_tokens.weight
-
-    assert cfg.d_mlp is not None  # keep mypy happy
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = qwen.model.layers[l].input_layernorm.weight
-
-        W_Q = qwen.model.layers[l].self_attn.q_proj.weight
-        W_K = qwen.model.layers[l].self_attn.k_proj.weight
-        W_V = qwen.model.layers[l].self_attn.v_proj.weight
-        W_Q = einops.rearrange(W_Q, "(n h) m->n m h", n=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(n h) m->n m h", n=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(n h) m->n m h", n=cfg.n_heads)
-
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        b_Q = qwen.model.layers[l].self_attn.q_proj.bias
-        b_Q = einops.rearrange(
-            b_Q,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-
-        b_K = qwen.model.layers[l].self_attn.k_proj.bias
-        b_K = einops.rearrange(
-            b_K,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-
-        b_V = qwen.model.layers[l].self_attn.v_proj.bias
-        b_V = einops.rearrange(
-            b_V,
-            "(n_head d_head) -> n_head d_head",
-            n_head=cfg.n_heads,
-        )
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = b_Q
-        state_dict[f"blocks.{l}.attn.b_K"] = b_K
-        state_dict[f"blocks.{l}.attn.b_V"] = b_V
-
-        W_O = qwen.model.layers[l].self_attn.o_proj.weight
-        W_O = einops.rearrange(W_O, "m (n h)->n h m", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.ln2.w"] = qwen.model.layers[l].post_attention_layernorm.weight
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = qwen.model.layers[l].mlp.up_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.W_gate"] = qwen.model.layers[l].mlp.gate_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = torch.zeros(cfg.d_mlp, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = qwen.model.layers[l].mlp.down_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-    state_dict["ln_final.w"] = qwen.model.norm.weight
-
-    state_dict["unembed.W_U"] = qwen.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-    return state_dict
-
-
-def convert_mistral_weights(mistral, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = mistral.model.embed_tokens.weight
-
-    assert cfg.n_key_value_heads is not None  # keep mypy happy
-    assert cfg.d_mlp is not None  # keep mypy happy
-
-    # Mistral has no biases anywhere
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = mistral.model.layers[l].input_layernorm.weight
-
-        W_Q = mistral.model.layers[l].self_attn.q_proj.weight
-        W_K = mistral.model.layers[l].self_attn.k_proj.weight
-        W_V = mistral.model.layers[l].self_attn.v_proj.weight
-        W_Q = einops.rearrange(W_Q, "(n h) m->n m h", n=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(n h) m->n m h", n=cfg.n_key_value_heads)
-        W_V = einops.rearrange(W_V, "(n h) m->n m h", n=cfg.n_key_value_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn._W_K"] = W_K
-        state_dict[f"blocks.{l}.attn._W_V"] = W_V
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn._b_K"] = torch.zeros(
-            cfg.n_key_value_heads, cfg.d_head, dtype=cfg.dtype
-        )
-        state_dict[f"blocks.{l}.attn._b_V"] = torch.zeros(
-            cfg.n_key_value_heads, cfg.d_head, dtype=cfg.dtype
-        )
-
-        W_O = mistral.model.layers[l].self_attn.o_proj.weight
-        W_O = einops.rearrange(W_O, "m (n h)->n h m", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.ln2.w"] = mistral.model.layers[l].post_attention_layernorm.weight
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = mistral.model.layers[l].mlp.up_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.W_gate"] = mistral.model.layers[l].mlp.gate_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = torch.zeros(cfg.d_mlp, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = mistral.model.layers[l].mlp.down_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-    state_dict["ln_final.w"] = mistral.model.norm.weight
-
-    state_dict["unembed.W_U"] = mistral.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-    return state_dict
-
-
-def convert_opt_weights(opt, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = opt.model.decoder.embed_tokens.weight
-    state_dict["pos_embed.W_pos"] = opt.model.decoder.embed_positions.weight[2:, :]
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = opt.model.decoder.layers[l].self_attn_layer_norm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = opt.model.decoder.layers[l].self_attn_layer_norm.bias
-
-        W_Q = opt.model.decoder.layers[l].self_attn.q_proj.weight
-        W_K = opt.model.decoder.layers[l].self_attn.k_proj.weight
-        W_V = opt.model.decoder.layers[l].self_attn.v_proj.weight
-        W_Q = einops.rearrange(
-            W_Q,
-            "(index d_head) d_model->index d_model d_head",
-            index=cfg.n_heads,
-        )
-        W_K = einops.rearrange(
-            W_K,
-            "(index d_head) d_model->index d_model d_head",
-            index=cfg.n_heads,
-        )
-        W_V = einops.rearrange(
-            W_V,
-            "(index d_head) d_model->index d_model d_head",
-            index=cfg.n_heads,
-        )
-
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        q_bias = einops.rearrange(
-            opt.model.decoder.layers[l].self_attn.q_proj.bias,
-            "(head_index d_head)->head_index d_head",
-            head_index=cfg.n_heads,
-            d_head=cfg.d_head,
-        )
-        k_bias = einops.rearrange(
-            opt.model.decoder.layers[l].self_attn.k_proj.bias,
-            "(head_index d_head)->head_index d_head",
-            head_index=cfg.n_heads,
-            d_head=cfg.d_head,
-        )
-        v_bias = einops.rearrange(
-            opt.model.decoder.layers[l].self_attn.v_proj.bias,
-            "(head_index d_head)->head_index d_head",
-            head_index=cfg.n_heads,
-            d_head=cfg.d_head,
-        )
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = q_bias
-        state_dict[f"blocks.{l}.attn.b_K"] = k_bias
-        state_dict[f"blocks.{l}.attn.b_V"] = v_bias
-
-        W_O = opt.model.decoder.layers[l].self_attn.out_proj.weight
-        W_O = einops.rearrange(
-            W_O,
-            "d_model (index d_head)->index d_head d_model",
-            index=cfg.n_heads,
-        )
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = opt.model.decoder.layers[l].self_attn.out_proj.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = opt.model.decoder.layers[l].final_layer_norm.weight
-        state_dict[f"blocks.{l}.ln2.b"] = opt.model.decoder.layers[l].final_layer_norm.bias
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = opt.model.decoder.layers[l].fc1.weight.T
-        state_dict[f"blocks.{l}.mlp.W_out"] = opt.model.decoder.layers[l].fc2.weight.T
-
-        state_dict[f"blocks.{l}.mlp.b_in"] = opt.model.decoder.layers[l].fc1.bias
-        state_dict[f"blocks.{l}.mlp.b_out"] = opt.model.decoder.layers[l].fc2.bias
-    state_dict["ln_final.w"] = opt.model.decoder.final_layer_norm.weight
-    state_dict["ln_final.b"] = opt.model.decoder.final_layer_norm.bias
-    state_dict["unembed.W_U"] = opt.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-    return state_dict
-
-
-def convert_neel_solu_old_weights(state_dict: dict, cfg: HookedTransformerConfig):
-    """
-    Converts the weights of my old SoLU models to the HookedTransformer format.
-    Takes as input a state dict, *not* a model object.
-
-    There are a bunch of dumb bugs in the original code, sorry!
-
-    Models 1L, 2L, 4L and 6L have left facing weights (ie, weights have shape
-    [dim_out, dim_in]) while HookedTransformer does right facing (ie [dim_in,
-    dim_out]).
-
-    8L has *just* a left facing W_pos, the rest right facing.
-
-    And some models were trained with
-    """
-    # Early models have left facing W_pos
-    reverse_pos = cfg.n_layers <= 8
-
-    # Models prior to 8L have left facing everything (8L has JUST left facing W_pos - sorry! Stupid bug)
-    reverse_weights = cfg.n_layers <= 6
-
-    new_state_dict = {}
-    for k, v in state_dict.items():
-        k = k.replace("norm", "ln")
-        if k.startswith("ln."):
-            k = k.replace("ln.", "ln_final.")
-        new_state_dict[k] = v
-
-    if reverse_pos:
-        new_state_dict["pos_embed.W_pos"] = new_state_dict["pos_embed.W_pos"].T
-    if reverse_weights:
-        for k, v in new_state_dict.items():
-            if "W_" in k and "W_pos" not in k:
-                new_state_dict[k] = v.transpose(-2, -1)
-    return new_state_dict
-
-
-def convert_mingpt_weights(old_state_dict, cfg: HookedTransformerConfig):
-    # mingpt (https://github.com/karpathy/minGPT) is mostly similar to GPT-2,
-    # but doesn't concat the QKV matrices.
-    state_dict = {}
-
-    state_dict["embed.W_E"] = old_state_dict["tok_emb.weight"]
-    state_dict["pos_embed.W_pos"] = old_state_dict["pos_emb"].squeeze()
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = old_state_dict[f"blocks.{l}.ln1.weight"]
-        state_dict[f"blocks.{l}.ln1.b"] = old_state_dict[f"blocks.{l}.ln1.bias"]
-
-        W_Q = old_state_dict[f"blocks.{l}.attn.query.weight"]
-        W_K = old_state_dict[f"blocks.{l}.attn.key.weight"]
-        W_V = old_state_dict[f"blocks.{l}.attn.value.weight"]
-        W_Q = einops.rearrange(W_Q, "(i h) m->i m h", i=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(i h) m->i m h", i=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(i h) m->i m h", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        q_bias = einops.rearrange(
-            old_state_dict[f"blocks.{l}.attn.query.bias"], "(i h)->i h", i=cfg.n_heads
-        )
-        k_bias = einops.rearrange(
-            old_state_dict[f"blocks.{l}.attn.key.bias"], "(i h)->i h", i=cfg.n_heads
-        )
-        v_bias = einops.rearrange(
-            old_state_dict[f"blocks.{l}.attn.value.bias"], "(i h)->i h", i=cfg.n_heads
-        )
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = q_bias
-        state_dict[f"blocks.{l}.attn.b_K"] = k_bias
-        state_dict[f"blocks.{l}.attn.b_V"] = v_bias
-
-        W_O = old_state_dict[f"blocks.{l}.attn.proj.weight"]
-        W_O = einops.rearrange(W_O, "m (i h)->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = old_state_dict[f"blocks.{l}.attn.proj.bias"]
-
-        state_dict[f"blocks.{l}.ln2.w"] = old_state_dict[f"blocks.{l}.ln2.weight"]
-        state_dict[f"blocks.{l}.ln2.b"] = old_state_dict[f"blocks.{l}.ln2.bias"]
-
-        W_in = old_state_dict[f"blocks.{l}.mlp.0.weight"]
-        state_dict[f"blocks.{l}.mlp.W_in"] = W_in.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = old_state_dict[f"blocks.{l}.mlp.0.bias"]
-
-        W_out = old_state_dict[f"blocks.{l}.mlp.2.weight"]
-        state_dict[f"blocks.{l}.mlp.W_out"] = W_out.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = old_state_dict[f"blocks.{l}.mlp.2.bias"]
-
-    state_dict["unembed.W_U"] = old_state_dict["head.weight"].T
-
-    state_dict["ln_final.w"] = old_state_dict["ln_f.weight"]
-    state_dict["ln_final.b"] = old_state_dict["ln_f.bias"]
-
-    return state_dict
-
-
-def convert_nanogpt_weights(old_state_dict, cfg: HookedTransformerConfig):
-    """For https://github.com/karpathy/nanoGPT
-    There are two complications with converting nanogpt models:
-    The first is that some state dicts have an unwanted prefix on keys that needs to be removed.
-    The second is that the models can be saved with or without bias. By default, there
-    is no bias. This function can handle both cases."""
-    # Nanogpt models saved after torch.compile() have this unwanted prefix
-    # This is a simple way to remove it
-    unwanted_prefix = "_orig_mod."
-    for k, v in list(old_state_dict.items()):
-        if k.startswith(unwanted_prefix):
-            old_state_dict[k[len(unwanted_prefix) :]] = old_state_dict.pop(k)
-
-    new_state_dict = {}
-    new_state_dict["pos_embed.W_pos"] = old_state_dict["transformer.wpe.weight"]
-    new_state_dict["embed.W_E"] = old_state_dict["transformer.wte.weight"]
-
-    new_state_dict["ln_final.w"] = old_state_dict["transformer.ln_f.weight"]
-    new_state_dict["ln_final.b"] = torch.zeros_like(old_state_dict["transformer.ln_f.weight"])
-    new_state_dict["unembed.W_U"] = old_state_dict["lm_head.weight"].T
-
-    bias = False
-    if "transformer.ln_f.bias" in old_state_dict:
-        bias = True
-        new_state_dict["ln_final.b"] = old_state_dict["transformer.ln_f.bias"]
-
-    for layer in range(cfg.n_layers):
-        layer_key = f"transformer.h.{layer}"
-
-        new_state_dict[f"blocks.{layer}.ln1.w"] = old_state_dict[f"{layer_key}.ln_1.weight"]
-        # A bias of zeros is required for folding layer norm
-        new_state_dict[f"blocks.{layer}.ln1.b"] = torch.zeros_like(
-            old_state_dict[f"{layer_key}.ln_1.weight"]
-        )
-        new_state_dict[f"blocks.{layer}.ln2.w"] = old_state_dict[f"{layer_key}.ln_2.weight"]
-        new_state_dict[f"blocks.{layer}.ln2.b"] = torch.zeros_like(
-            old_state_dict[f"{layer_key}.ln_2.weight"]
-        )
-
-        W = old_state_dict[f"{layer_key}.attn.c_attn.weight"]
-        W_Q, W_K, W_V = torch.tensor_split(W, 3, dim=0)
-        W_Q = einops.rearrange(W_Q, "(i h) m->i m h", i=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(i h) m->i m h", i=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "(i h) m->i m h", i=cfg.n_heads)
-        new_state_dict[f"blocks.{layer}.attn.W_Q"] = W_Q
-        new_state_dict[f"blocks.{layer}.attn.W_K"] = W_K
-        new_state_dict[f"blocks.{layer}.attn.W_V"] = W_V
-
-        W_O = old_state_dict[f"{layer_key}.attn.c_proj.weight"]
-        W_O = einops.rearrange(W_O, "m (i h)->i h m", i=cfg.n_heads)
-        new_state_dict[f"blocks.{layer}.attn.W_O"] = W_O
-
-        new_state_dict[f"blocks.{layer}.mlp.W_in"] = old_state_dict[
-            f"{layer_key}.mlp.c_fc.weight"
-        ].T
-        new_state_dict[f"blocks.{layer}.mlp.W_out"] = old_state_dict[
-            f"{layer_key}.mlp.c_proj.weight"
-        ].T
-
-        if bias:
-            new_state_dict[f"blocks.{layer}.ln1.b"] = old_state_dict[f"{layer_key}.ln_1.bias"]
-            new_state_dict[f"blocks.{layer}.ln2.b"] = old_state_dict[f"{layer_key}.ln_2.bias"]
-            new_state_dict[f"blocks.{layer}.mlp.b_in"] = old_state_dict[
-                f"{layer_key}.mlp.c_fc.bias"
-            ]
-            new_state_dict[f"blocks.{layer}.mlp.b_out"] = old_state_dict[
-                f"{layer_key}.mlp.c_proj.bias"
-            ]
-
-            B = old_state_dict[f"{layer_key}.attn.c_attn.bias"]
-            B_Q, B_K, B_V = torch.tensor_split(B, 3, dim=0)
-            B_Q = einops.rearrange(B_Q, "(i h)->i h", i=cfg.n_heads)
-            B_K = einops.rearrange(B_K, "(i h)->i h", i=cfg.n_heads)
-            B_V = einops.rearrange(B_V, "(i h)->i h", i=cfg.n_heads)
-            new_state_dict[f"blocks.{layer}.attn.b_Q"] = B_Q
-            new_state_dict[f"blocks.{layer}.attn.b_K"] = B_K
-            new_state_dict[f"blocks.{layer}.attn.b_V"] = B_V
-            new_state_dict[f"blocks.{layer}.attn.b_O"] = old_state_dict[
-                f"{layer_key}.attn.c_proj.bias"
-            ]
-
-    return new_state_dict
-
-
-def convert_bert_weights(bert, cfg: HookedTransformerConfig):
-    embeddings = bert.bert.embeddings
-    state_dict = {
-        "embed.embed.W_E": embeddings.word_embeddings.weight,
-        "embed.pos_embed.W_pos": embeddings.position_embeddings.weight,
-        "embed.token_type_embed.W_token_type": embeddings.token_type_embeddings.weight,
-        "embed.ln.w": embeddings.LayerNorm.weight,
-        "embed.ln.b": embeddings.LayerNorm.bias,
-    }
-
-    for l in range(cfg.n_layers):
-        block = bert.bert.encoder.layer[l]
-        state_dict[f"blocks.{l}.attn.W_Q"] = einops.rearrange(
-            block.attention.self.query.weight, "(i h) m -> i m h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.b_Q"] = einops.rearrange(
-            block.attention.self.query.bias, "(i h) -> i h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.W_K"] = einops.rearrange(
-            block.attention.self.key.weight, "(i h) m -> i m h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.b_K"] = einops.rearrange(
-            block.attention.self.key.bias, "(i h) -> i h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.W_V"] = einops.rearrange(
-            block.attention.self.value.weight, "(i h) m -> i m h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.b_V"] = einops.rearrange(
-            block.attention.self.value.bias, "(i h) -> i h", i=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.W_O"] = einops.rearrange(
-            block.attention.output.dense.weight,
-            "m (i h) -> i h m",
-            i=cfg.n_heads,
-        )
-        state_dict[f"blocks.{l}.attn.b_O"] = block.attention.output.dense.bias
-        state_dict[f"blocks.{l}.ln1.w"] = block.attention.output.LayerNorm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = block.attention.output.LayerNorm.bias
-        state_dict[f"blocks.{l}.mlp.W_in"] = einops.rearrange(
-            block.intermediate.dense.weight, "mlp model -> model mlp"
-        )
-        state_dict[f"blocks.{l}.mlp.b_in"] = block.intermediate.dense.bias
-        state_dict[f"blocks.{l}.mlp.W_out"] = einops.rearrange(
-            block.output.dense.weight, "model mlp -> mlp model"
-        )
-        state_dict[f"blocks.{l}.mlp.b_out"] = block.output.dense.bias
-        state_dict[f"blocks.{l}.ln2.w"] = block.output.LayerNorm.weight
-        state_dict[f"blocks.{l}.ln2.b"] = block.output.LayerNorm.bias
-
-    mlm_head = bert.cls.predictions
-    state_dict["mlm_head.W"] = mlm_head.transform.dense.weight
-    state_dict["mlm_head.b"] = mlm_head.transform.dense.bias
-    state_dict["mlm_head.ln.w"] = mlm_head.transform.LayerNorm.weight
-    state_dict["mlm_head.ln.b"] = mlm_head.transform.LayerNorm.bias
-    # Note: BERT uses tied embeddings
-    state_dict["unembed.W_U"] = embeddings.word_embeddings.weight.T
-    # "unembed.W_U": mlm_head.decoder.weight.T,
-    state_dict["unembed.b_U"] = mlm_head.bias
-
-    return state_dict
-
-
-def convert_bloom_weights(bloom, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = bloom.transformer.word_embeddings.weight
-
-    # Bloom uses post embedding layer norm
-    state_dict["embed.ln.w"] = bloom.transformer.word_embeddings_layernorm.weight
-    state_dict["embed.ln.b"] = bloom.transformer.word_embeddings_layernorm.bias
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = bloom.transformer.h[l].input_layernorm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = bloom.transformer.h[l].input_layernorm.bias
-
-        W = bloom.transformer.h[l].self_attention.query_key_value.weight
-
-        W_split = W.T.reshape(cfg.d_model, cfg.n_heads, 3, cfg.d_head)
-
-        W_Q, W_K, W_V = W_split[..., 0, :], W_split[..., 1, :], W_split[..., 2, :]
-        W_Q = einops.rearrange(W_Q, "m n h ->n m h", n=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "m n h ->n m h", n=cfg.n_heads)
-        W_V = einops.rearrange(W_V, "m n h ->n m h", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        qkv_bias = bloom.transformer.h[l].self_attention.query_key_value.bias
-        qkv_bias = qkv_bias.reshape(cfg.n_heads, 3, cfg.d_head)
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = qkv_bias[:, 0, :]
-        state_dict[f"blocks.{l}.attn.b_K"] = qkv_bias[:, 1, :]
-        state_dict[f"blocks.{l}.attn.b_V"] = qkv_bias[:, 2, :]
-
-        W_O = bloom.transformer.h[l].self_attention.dense.weight.T  # [1024, 1024]
-        W_O = einops.rearrange(W_O, "(n h) m->n h m", n=cfg.n_heads)  # [n_heads, d_head, d_model]
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = bloom.transformer.h[l].self_attention.dense.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = bloom.transformer.h[l].post_attention_layernorm.weight
-        state_dict[f"blocks.{l}.ln2.b"] = bloom.transformer.h[l].post_attention_layernorm.bias
-
-        W_in = bloom.transformer.h[l].mlp.dense_h_to_4h.weight.T
-        state_dict[f"blocks.{l}.mlp.W_in"] = W_in
-        state_dict[f"blocks.{l}.mlp.b_in"] = bloom.transformer.h[l].mlp.dense_h_to_4h.bias
-
-        W_out = bloom.transformer.h[l].mlp.dense_4h_to_h.weight.T
-        state_dict[f"blocks.{l}.mlp.W_out"] = W_out
-        state_dict[f"blocks.{l}.mlp.b_out"] = bloom.transformer.h[l].mlp.dense_4h_to_h.bias
-    state_dict["unembed.W_U"] = bloom.lm_head.weight.T
-
-    state_dict["ln_final.w"] = bloom.transformer.ln_f.weight
-    state_dict["ln_final.b"] = bloom.transformer.ln_f.bias
-    return state_dict
-
-
-def convert_coder_weights(model, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = model.transformer.wte.weight
-    state_dict["pos_embed.W_pos"] = model.transformer.wpe.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = model.transformer.h[l].ln_1.weight
-        state_dict[f"blocks.{l}.ln1.b"] = model.transformer.h[l].ln_1.bias
-
-        # In GPT-2, q,k,v are produced by one big linear map, whose output is
-        # concat([q, k, v])
-        W_KV = model.transformer.h[l].attn.kv_attn.weight  # [d_model, 2 * d_head]
-        W_K, W_V = torch.tensor_split(W_KV, 2, dim=1)
-        W_Q = model.transformer.h[l].attn.q_attn.weight  # [d_model, d_model]
-        W_Q = einops.rearrange(W_Q, "m (i h)->i m h", i=cfg.n_heads)
-        W_K = einops.repeat(W_K, "m h -> i m h", i=cfg.n_heads)
-        W_V = einops.repeat(W_V, "m h -> i m h", i=cfg.n_heads)
-
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        b_Q = einops.rearrange(
-            model.transformer.h[l].attn.q_attn.bias,
-            "(index head)-> index head",
-            index=cfg.n_heads,
-            head=cfg.d_head,
-        )
-        b_KV = model.transformer.h[l].attn.kv_attn.bias  # [2 * d_head]
-        b_K, b_V = torch.tensor_split(b_KV, 2, dim=0)
-        b_K = einops.repeat(b_K, "head -> index head", index=cfg.n_heads)
-        b_V = einops.repeat(b_V, "head -> index head", index=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.b_Q"] = b_Q
-        state_dict[f"blocks.{l}.attn.b_K"] = b_K
-        state_dict[f"blocks.{l}.attn.b_V"] = b_V
-
-        W_O = model.transformer.h[l].attn.c_proj.weight
-        W_O = einops.rearrange(W_O, "(i h) m->i h m", i=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = model.transformer.h[l].attn.c_proj.bias
-
-        state_dict[f"blocks.{l}.ln2.w"] = model.transformer.h[l].ln_2.weight
-        state_dict[f"blocks.{l}.ln2.b"] = model.transformer.h[l].ln_2.bias
-
-        W_in = model.transformer.h[l].mlp.c_fc.weight
-        state_dict[f"blocks.{l}.mlp.W_in"] = W_in
-        state_dict[f"blocks.{l}.mlp.b_in"] = model.transformer.h[l].mlp.c_fc.bias
-
-        W_out = model.transformer.h[l].mlp.c_proj.weight
-        state_dict[f"blocks.{l}.mlp.W_out"] = W_out
-        state_dict[f"blocks.{l}.mlp.b_out"] = model.transformer.h[l].mlp.c_proj.bias
-    state_dict["unembed.W_U"] = model.lm_head.weight.T
-
-    state_dict["ln_final.w"] = model.transformer.ln_f.weight
-    state_dict["ln_final.b"] = model.transformer.ln_f.bias
-    return state_dict
-
-
-def convert_phi_weights(phi, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = phi.model.embed_tokens.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = phi.model.layers[l].input_layernorm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = phi.model.layers[l].input_layernorm.bias
-
-        W_Q = phi.model.layers[l].self_attn.q_proj.weight
-        W_K = phi.model.layers[l].self_attn.k_proj.weight
-        W_V = phi.model.layers[l].self_attn.v_proj.weight
-        W_Q = einops.rearrange(
-            W_Q, "(n_head d_head) d_model -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        W_K = einops.rearrange(
-            W_K, "(n_head d_head) d_model  -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        W_V = einops.rearrange(
-            W_V, "(n_head d_head) d_model  -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-
-        b_Q = phi.model.layers[l].self_attn.q_proj.bias
-        b_K = phi.model.layers[l].self_attn.k_proj.bias
-        b_V = phi.model.layers[l].self_attn.v_proj.bias
-        b_Q = einops.rearrange(b_Q, "(n_head d_head) -> n_head d_head", n_head=cfg.n_heads)
-        b_K = einops.rearrange(b_K, "(n_head d_head) -> n_head d_head", n_head=cfg.n_heads)
-        b_V = einops.rearrange(b_V, "(n_head d_head) -> n_head d_head", n_head=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.b_Q"] = b_Q
-        state_dict[f"blocks.{l}.attn.b_K"] = b_K
-        state_dict[f"blocks.{l}.attn.b_V"] = b_V
-
-        W_O = phi.model.layers[l].self_attn.dense.weight
-        W_O = einops.rearrange(
-            W_O, "d_model (n_head d_head) -> n_head d_head d_model", n_head=cfg.n_heads
-        )
-
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = phi.model.layers[l].self_attn.dense.bias
-
-        # Layer Norm 1 and 2 are tied.
-        state_dict[f"blocks.{l}.ln2.w"] = state_dict[f"blocks.{l}.ln1.w"]
-        state_dict[f"blocks.{l}.ln2.b"] = state_dict[f"blocks.{l}.ln1.b"]
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = phi.model.layers[l].mlp.fc1.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = phi.model.layers[l].mlp.fc1.bias
-        state_dict[f"blocks.{l}.mlp.W_out"] = phi.model.layers[l].mlp.fc2.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = phi.model.layers[l].mlp.fc2.bias
-
-    state_dict["ln_final.w"] = phi.model.final_layernorm.weight
-    state_dict["ln_final.b"] = phi.model.final_layernorm.bias
-
-    state_dict["unembed.W_U"] = phi.lm_head.weight.T
-    state_dict["unembed.b_U"] = phi.lm_head.bias
-
-    return state_dict
-
-
-def convert_phi3_weights(phi, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    state_dict["embed.W_E"] = phi.model.embed_tokens.weight
-
-    for l in range(cfg.n_layers):
-        state_dict[f"blocks.{l}.ln1.w"] = phi.model.layers[l].input_layernorm.weight
-        state_dict[f"blocks.{l}.ln1.b"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-        W = phi.model.layers[l].self_attn.qkv_proj.weight
-        W_Q, W_K, W_V = torch.tensor_split(W, 3, dim=0)
-        W_Q = einops.rearrange(
-            W_Q, "(n_head d_head) d_model -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        W_K = einops.rearrange(
-            W_K, "(n_head d_head) d_model  -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        W_V = einops.rearrange(
-            W_V, "(n_head d_head) d_model  -> n_head d_model d_head", n_head=cfg.n_heads
-        )
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.W_K"] = W_K
-        state_dict[f"blocks.{l}.attn.b_K"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn.W_V"] = W_V
-        state_dict[f"blocks.{l}.attn.b_V"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-
-        W_O = phi.model.layers[l].self_attn.o_proj.weight
-        W_O = einops.rearrange(
-            W_O, "d_model (n_head d_head) -> n_head d_head d_model", n_head=cfg.n_heads
-        )
-
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.ln2.w"] = phi.model.layers[l].post_attention_layernorm.weight
-        state_dict[f"blocks.{l}.ln2.b"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-        W = phi.model.layers[l].mlp.gate_up_proj.weight.T
-        W_gate, W_in = torch.tensor_split(W, 2, dim=1)
-        state_dict[f"blocks.{l}.mlp.W_in"] = W_in
-        state_dict[f"blocks.{l}.mlp.W_gate"] = W_gate
-        state_dict[f"blocks.{l}.mlp.W_out"] = phi.model.layers[l].mlp.down_proj.weight.T
-
-    state_dict["ln_final.w"] = phi.model.norm.weight
-
-    state_dict["unembed.W_U"] = phi.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
-    return state_dict
-
-
-def convert_gemma_weights(gemma, cfg: HookedTransformerConfig):
-    state_dict = {}
-
-    assert cfg.n_key_value_heads is not None  # keep mypy happy
-    assert cfg.d_mlp is not None  # keep mypy happy
-
-    # Gemma Models scale embeddings by multiplying by sqrt(d_model), use hidden state type to match
-    # HF implementation
-    state_dict["embed.W_E"] = gemma.model.embed_tokens.weight * torch.tensor(
-        cfg.d_model**0.5, dtype=cfg.dtype
-    )
-
-    # Gemma has no biases anywhere
-    for l in range(cfg.n_layers):
-        # GemmaRMSNorm adds 1 to weights before multiplying by input, keep RMS calcs in float32
-        state_dict[f"blocks.{l}.ln1.w"] = gemma.model.layers[
-            l
-        ].input_layernorm.weight.float() + torch.ones_like(
-            gemma.model.layers[l].input_layernorm.weight, dtype=torch.float32
-        )
-
-        W_Q = gemma.model.layers[l].self_attn.q_proj.weight
-        W_K = gemma.model.layers[l].self_attn.k_proj.weight
-        W_V = gemma.model.layers[l].self_attn.v_proj.weight
-        W_Q = einops.rearrange(W_Q, "(n h) m->n m h", n=cfg.n_heads)
-        W_K = einops.rearrange(W_K, "(n h) m->n m h", n=cfg.n_key_value_heads)
-        W_V = einops.rearrange(W_V, "(n h) m->n m h", n=cfg.n_key_value_heads)
-        state_dict[f"blocks.{l}.attn.W_Q"] = W_Q
-        state_dict[f"blocks.{l}.attn._W_K"] = W_K
-        state_dict[f"blocks.{l}.attn._W_V"] = W_V
-
-        state_dict[f"blocks.{l}.attn.b_Q"] = torch.zeros(cfg.n_heads, cfg.d_head, dtype=cfg.dtype)
-        state_dict[f"blocks.{l}.attn._b_K"] = torch.zeros(
-            cfg.n_key_value_heads, cfg.d_head, dtype=cfg.dtype
-        )
-        state_dict[f"blocks.{l}.attn._b_V"] = torch.zeros(
-            cfg.n_key_value_heads, cfg.d_head, dtype=cfg.dtype
-        )
-
-        W_O = gemma.model.layers[l].self_attn.o_proj.weight
-        W_O = einops.rearrange(W_O, "m (n h)->n h m", n=cfg.n_heads)
-        state_dict[f"blocks.{l}.attn.W_O"] = W_O
-
-        state_dict[f"blocks.{l}.attn.b_O"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-        # GemmaRMSNorm adds 1 to weights before multiplying by input, keep RMS calcs in float32
-        state_dict[f"blocks.{l}.ln2.w"] = gemma.model.layers[
-            l
-        ].post_attention_layernorm.weight.float() + torch.ones_like(
-            gemma.model.norm.weight, dtype=torch.float32
-        )
-
-        state_dict[f"blocks.{l}.mlp.W_in"] = gemma.model.layers[l].mlp.up_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.W_gate"] = gemma.model.layers[l].mlp.gate_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_in"] = torch.zeros(cfg.d_mlp, dtype=cfg.dtype)
-
-        state_dict[f"blocks.{l}.mlp.W_out"] = gemma.model.layers[l].mlp.down_proj.weight.T
-        state_dict[f"blocks.{l}.mlp.b_out"] = torch.zeros(cfg.d_model, dtype=cfg.dtype)
-
-    # GemmaRMSNorm adds 1 to weights before multiplying by input, keep RMS calcs in float32
-    state_dict["ln_final.w"] = gemma.model.norm.weight.float() + torch.ones_like(
-        gemma.model.norm.weight, dtype=torch.float32
-    )
-
-    state_dict["unembed.W_U"] = gemma.lm_head.weight.T
-    state_dict["unembed.b_U"] = torch.zeros(cfg.d_vocab, dtype=cfg.dtype)
-
     return state_dict
