+++ transformer_lens/pretrained/weight_conversions/__init__.py
@@ -1,1 +1,19 @@
 from .mixtral import convert_mixtral_weights
+from .neo import convert_neo_weights
+from .gpt2 import convert_gpt2_weights
+from .opt import convert_opt_weights
+from .gptj import convert_gptj_weights
+from .neox import convert_neox_weights
+from .llama import convert_llama_weights
+from .bert import convert_bert_weights
+from .mistral import convert_mistral_weights
+from .bloom import convert_bloom_weights
+from .coder import convert_coder_weights
+from .qwen import convert_qwen_weights
+from .qwen2 import convert_qwen2_weights
+from .phi import convert_phi_weights
+from .phi3 import convert_phi3_weights
+from .gemma import convert_gemma_weights
+from .mingpt import convert_mingpt_weights
+from .nanogpt import convert_nanogpt_weights
+from .neel_solu_old import convert_neel_solu_old_weights
