+++ transformer_lens/pretrained/weight_conversions/neel_solu_old.py
@@ -1,0 +1,38 @@
+from transformer_lens.HookedTransformerConfig import HookedTransformerConfig
+
+
+def convert_neel_solu_old_weights(state_dict: dict, cfg: HookedTransformerConfig):
+    """
+    Converts the weights of my old SoLU models to the HookedTransformer format.
+    Takes as input a state dict, *not* a model object.
+
+    There are a bunch of dumb bugs in the original code, sorry!
+
+    Models 1L, 2L, 4L and 6L have left facing weights (ie, weights have shape
+    [dim_out, dim_in]) while HookedTransformer does right facing (ie [dim_in,
+    dim_out]).
+
+    8L has *just* a left facing W_pos, the rest right facing.
+
+    And some models were trained with
+    """
+    # Early models have left facing W_pos
+    reverse_pos = cfg.n_layers <= 8
+
+    # Models prior to 8L have left facing everything (8L has JUST left facing W_pos - sorry! Stupid bug)
+    reverse_weights = cfg.n_layers <= 6
+
+    new_state_dict = {}
+    for k, v in state_dict.items():
+        k = k.replace("norm", "ln")
+        if k.startswith("ln."):
+            k = k.replace("ln.", "ln_final.")
+        new_state_dict[k] = v
+
+    if reverse_pos:
+        new_state_dict["pos_embed.W_pos"] = new_state_dict["pos_embed.W_pos"].T
+    if reverse_weights:
+        for k, v in new_state_dict.items():
+            if "W_" in k and "W_pos" not in k:
+                new_state_dict[k] = v.transpose(-2, -1)
+    return new_state_dict
